+++ domain/model/src/main/java/co/com/crediya/model/user/User.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @NoArgsConstructor
@@ -15,7 +17,7 @@ public class User {
     private String rol;
     private String name;
     private String lastName;
-    private Double baseSalary;
+    private BigDecimal baseSalary;
 
     public static User fromJwt(String email, String rol) {
         return new User(email, rol, null, null, null);
@@ -25,7 +27,7 @@ public class User {
         return new User(email, null, null, null, null);
     }
 
-    public static User fromExtraInfo(String name, String lastName, Double baseSalary) {
+    public static User fromExtraInfo(String name, String lastName, BigDecimal baseSalary) {
         return new User(null, null, name, lastName, baseSalary);
     }
 }
